5. Конвертер валюты [#41583 #158650]

+++ src/ru/job4j/ru/job4j/Converter.java
@@ -3,27 +3,27 @@ package ru.job4j.ru.job4j;
 public class Converter {
     public static float rubleToEuro(float value) {
         float rsl = value / 70;
-        return (float) rsl;
+        return rsl;
     }
     public static float rubleToDollar(float value) {
         float rsl = value / 60;
-        return (float) rsl;
+        return rsl;
     }
     public static float euroToDollar(float value) {
         float rsl = value * 70 / 60;
-        return (float) rsl;
+        return rsl;
     }
     public static float euroToRuble(float value) {
         float rsl = value * 70;
-        return (float) rsl;
+        return rsl;
     }
     public static float dollarToEuro(float value) {
         float rsl = value * 60 / 70;
-        return (float) rsl;
+        return rsl;
     }
     public static float dollarToRuble(float value) {
         float rsl = value * 60;
-        return (float) rsl;
+        return rsl;
     }
     public static void main(String[] args) {
         float euro = Converter.rubleToEuro(140);
